Move insects into their own module
