+++ src/Modules/Effects/_Enums.cs
@@ -13,59 +13,6 @@ public static class _Enums
 	public static RoomSettings.RoomEffect.Type FogOfWarDarkened = new(nameof(FogOfWarDarkened), false); //disabled
 
 	/// <summary>
-	/// Glowing Swimmer effect enum
-	/// </summary>
-	public static RoomSettings.RoomEffect.Type GlowingSwimmers = new(nameof(GlowingSwimmers), true);
-	/// <summary>
-	/// Colored Camo Beetle effect enum
-	/// </summary>
-	public static RoomSettings.RoomEffect.Type ColoredCamoBeetles = new(nameof(ColoredCamoBeetles), true);
-	/// <summary>
-	/// Mosquito Insect effect enum
-	/// </summary>
-	public static RoomSettings.RoomEffect.Type MosquitoInsects = new(nameof(MosquitoInsects), true);
-	/// Butterfly A effect enum
-	/// </summary>
-	public static RoomSettings.RoomEffect.Type ButterfliesA = new(nameof(ButterfliesA), true);
-	/// <summary>
-	/// Butterfly B effect enum
-	/// </summary>
-	public static RoomSettings.RoomEffect.Type ButterfliesB = new(nameof(ButterfliesB), true);
-	/// <summary>
-	/// Zipper effect enum
-	/// </summary>
-	public static RoomSettings.RoomEffect.Type Zippers = new(nameof(Zippers), true);
-	
-	/// <summary>
-	/// Glowing Swimmer insect enum
-	/// </summary>
-	public static CosmeticInsect.Type GlowingSwimmerInsect = new(nameof(GlowingSwimmerInsect), true);
-	/// <summary>
-	/// Colored Camo Beetle insect enum
-	/// </summary>
-	public static CosmeticInsect.Type ColoredCamoBeetle = new(nameof(ColoredCamoBeetle), true);
-	/// <summary>
-	/// Mosquito Insect enum
-	/// </summary>
-	public static CosmeticInsect.Type MosquitoInsect = new(nameof(MosquitoInsect), true);
-	/// <summary>
-	/// Butterfly A enum
-	/// </summary>
-	public static CosmeticInsect.Type ButterflyA = new(nameof(ButterflyA), true);
-	/// <summary>
-	/// Butterfly B enum
-	/// </summary>
-	public static CosmeticInsect.Type ButterflyB = new(nameof(ButterflyB), true);
-	/// <summary>
-	/// Zipper enum
-	/// </summary>
-	public static CosmeticInsect.Type Zipper = new(nameof(Zipper), true);
-	/// <summary>
-	/// Circuit fly enum
-	/// </summary>
-	public static CosmeticInsect.Type CircuitFly = new(nameof(CircuitFly), true);
-
-	/// <summary>
 	/// ReplaceEffectColorA
 	/// </summary>
 	public static RoomSettings.RoomEffect.Type ReplaceEffectColorA = new(nameof(ReplaceEffectColorA), true);
