+++ src/Modules/Effects/_Module.cs
@@ -37,12 +37,6 @@ public static class _Module
 
 	internal static void Enable()
 	{
-		GlowingSwimmersCI.Apply();
-		ColoredCamoBeetlesCI.Apply();
-		MosquitoInsectsCI.Apply();
-		ButterfliesCI.Apply();
-		ZippersCI.Apply();
-		
 		ColorRoomEffect.Apply();
 		ReplaceEffectColor.Apply();
 		HiveColorAlpha.Apply();
@@ -62,12 +56,6 @@ public static class _Module
 
 	internal static void Disable()
 	{
-		GlowingSwimmersCI.Undo();
-		ColoredCamoBeetlesCI.Undo();
-		MosquitoInsectsCI.Undo();
-		ButterfliesCI.Undo();
-		ZippersCI.Undo();
-
 		ColorRoomEffect.Undo();
 		ReplaceEffectColor.Undo();
 		HiveColorAlpha.Undo();
@@ -89,13 +77,6 @@ public static class _Module
 			type == _Enums.ReplaceEffectColorB ||
 			type == _Enums.FogOfWarSolid ||
 			type == _Enums.FogOfWarDarkened ||
-
-			type == _Enums.GlowingSwimmers ||
-			type == _Enums.ColoredCamoBeetles ||
-			type == _Enums.MosquitoInsects ||
-			type == _Enums.ButterfliesA ||
-			type == _Enums.ButterfliesB ||
-			type == _Enums.Zippers ||
 
 			type == _Enums.PWMalfunction ||
 			type == _Enums.HiveColorAlpha ||
