+++ src/Modules/Insects/ColoredCamoBeetlesCI.cs
@@ -1,6 +1,6 @@
 ﻿using System.Runtime.CompilerServices;
 
-namespace RegionKit.Modules.Effects;
+namespace RegionKit.Modules.Insects;
 
 /// <summary>
 /// By LB/M4rbleL1ne
