+++ src/Modules/Insects/_Enums.cs
@@ -1,0 +1,59 @@
+﻿namespace RegionKit.Modules.Insects
+{
+	public static class _Enums
+	{
+		/// <summary>
+		/// Glowing Swimmer effect enum
+		/// </summary>
+		public static RoomSettings.RoomEffect.Type GlowingSwimmers = new(nameof(GlowingSwimmers), true);
+		/// <summary>
+		/// Colored Camo Beetle effect enum
+		/// </summary>
+		public static RoomSettings.RoomEffect.Type ColoredCamoBeetles = new(nameof(ColoredCamoBeetles), true);
+		/// <summary>
+		/// Mosquito Insect effect enum
+		/// </summary>
+		public static RoomSettings.RoomEffect.Type MosquitoInsects = new(nameof(MosquitoInsects), true);
+		/// Butterfly A effect enum
+		/// </summary>
+		public static RoomSettings.RoomEffect.Type ButterfliesA = new(nameof(ButterfliesA), true);
+		/// <summary>
+		/// Butterfly B effect enum
+		/// </summary>
+		public static RoomSettings.RoomEffect.Type ButterfliesB = new(nameof(ButterfliesB), true);
+		/// <summary>
+		/// Zipper effect enum
+		/// </summary>
+		public static RoomSettings.RoomEffect.Type Zippers = new(nameof(Zippers), true);
+
+		/// <summary>
+		/// Glowing Swimmer insect enum
+		/// </summary>
+		public static CosmeticInsect.Type GlowingSwimmerInsect = new(nameof(GlowingSwimmerInsect), true);
+		/// <summary>
+		/// Colored Camo Beetle insect enum
+		/// </summary>
+		public static CosmeticInsect.Type ColoredCamoBeetle = new(nameof(ColoredCamoBeetle), true);
+		/// <summary>
+		/// Mosquito Insect enum
+		/// </summary>
+		public static CosmeticInsect.Type MosquitoInsect = new(nameof(MosquitoInsect), true);
+		/// <summary>
+		/// Butterfly A enum
+		/// </summary>
+		public static CosmeticInsect.Type ButterflyA = new(nameof(ButterflyA), true);
+		/// <summary>
+		/// Butterfly B enum
+		/// </summary>
+		public static CosmeticInsect.Type ButterflyB = new(nameof(ButterflyB), true);
+		/// <summary>
+		/// Zipper enum
+		/// </summary>
+		public static CosmeticInsect.Type Zipper = new(nameof(Zipper), true);
+
+		/// <summary>
+		/// Insect category
+		/// </summary>
+		public static DevInterface.RoomSettingsPage.DevEffectsCategories RegionKit = new(nameof(RegionKit), true);
+	}
+}
